test: migrate DataUseAlertBox to vitest
